更新了所有dao层接口，优化了同步问题

+++ src/main/java/com/jxust/qq/superquestionlib/dao/mapper/admin/repository/EsQuestionRepository.java
@@ -3,6 +3,6 @@ package com.jxust.qq.superquestionlib.dao.mapper.admin.repository;
 import com.jxust.qq.superquestionlib.dto.admin.EsQuestion;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface EsQuestionRepository extends ElasticsearchRepository<EsQuestion,Integer>
+public interface EsQuestionRepository extends ElasticsearchRepository<EsQuestion,Long>
 {
 }
